+++ packages/apps/bridges/crabDVM-heco/utils/index.ts
@@ -1,2 +1,1 @@
 export * from './tx';
-export * from './fee';
