+++ packages/shared/config/network/heco.ts
@@ -12,7 +12,7 @@ export const hecoConfig: EthereumChainConfig = {
   isTest: false,
   logos: [{ name: 'heco.png', type: 'main' }],
   name: 'heco',
-  provider: 'https://http-mainnet.hecochain.com',
+  provider: 'wss://ws-mainnet.hecochain.com',
   social: {
     portal: 'https://www.hecochain.com/en-us/',
     github: 'https://github.com/stars-labs/',
